refactor: attach socket.io to express server
