+++ app/routes/index.js
@@ -5,7 +5,7 @@ const router = express.Router();
 const fs = require('fs-extra');
 const path = require('path');
 const { v4: uuidv4 } = require('uuid');
-const { socket } = require('../socket/socket.js');
+const { socket } = require('../socket/client.js');
 const { uploadImage } = require('../services/multer');
 const { zipFolderSync } = require('../utils/zip.js');
